add file: mutt-msg-sender-and-local-datetime2
The 'mutt-msg-sender-and-local-datetime2' program is a rewrite (in Go
(a.k.a. "golong")) of the 'mutt-msg-sender-and-local-datetime' program (which
was implemented as a bash script), as both an exercise in learning the Go
programming language and as an attempt to get better performance of an
interactive mutt session (which invokes the program once for every message in
a given mailbox when in the mutt 'index' view; that amounts to a noticible lag
(a second or so, with the original bash script version) when invoked hundreds
of times in quick succession when using a large display).

The compiled version of this program is intended to be invoked from mutt at
runtime (via a format string that ends with a '|' character; see the section
"Filters" in the mutt manual. The output from this program will be used by
mutt as the updated filter string, which mutt will parse again.