add file: src/main/bash/bin/ads-mutt-index-format-custom
This is the bash script in which I've prototyped my most recent "new best way"
to configure the mutt index view. It includes my timestamp hacks from a few
years back (to conditionally show the sender's timestamp if it is in a
timezone different from my own, in a column in the mutt index view dedicated
for the purpose), and also includes a new hack to conditionally display a the
name of a known, subscribed-to mailing list (also in a column dedicated for
the purpose), if that is the origin of the message.

Fundamentally, when scanning the mutt index view, I want to be able to see at
a glance which messages are from a given mailing list, yet at the same time I
want to be able to see the sender's email address or name, including my own
when I submit posts. Once upon a time, mailing lists would often prefix the
list subject line with the list name, but these days that breaks DKIM
processing so many lists have moved away from it.

Mutt has the '%L' expansion in the 'index_format' "format string", but note
that it provides one behavior or the other, whereas I want both at the same
time.

The behavior offered by this mutt filter is superior to that anyway, IMO,
because the column of list names creates blocks and gaps in the display that
make the entire thing easier to interpret at a glance and to otherwise scan
quickly. In threaded view, the labelling of consecutive messages for a given
list make it easy both to hone in on such threads or skip over them
quickly. Similarly, the gaps in that same column for non-list messages also
makes it very easy to hone in on- or quickly skip over them, as well.

Aside from the fact that this program is very tightly coupled to my mutt
configuration, the most hackish thing about it is the hard-coded
'gl_ignorable_nonlists' array with the name(s) of non-list account names that
should not produce any value in the new column of output dedicated to the
mailing list name. There's a FIXME comment in the code to maybe pull that from
a configuration file or something. It may even be possible to have mutt pass
the value along somehow from its own configuration. At the moment, the list
contains a single value: 'ads'.

Based on txt messaging correspondence with friends, I wrote this program
sometime within the first two week of January 2019.