add file: src/main/golang/bin/ads-mutt-index-format-custom2.go
This is the golang rewrite of the 'ads-mutt-index-format-custom' bash program.

In terms of functionality, the two programs should be nearly identical, but
the golang program executes more quickly.

However, in it's current form, it is still pretty far off the mark from what I
was hoping for. When invoked by mutt back-to-back, say, 100 times, the
processing when done with the bash version of the program was taking between 5
and 6 seconds total, but the golang version is taking between 4 and 5 seconds.