Add dep on 'unicode-segmentation' crate, version 1.2.0
Context path:

    src/main/rust/Cargo.toml

The Rust standard library has fine support for Unicode, and in particular
UTF-8 encoding (so can handle byte sequences of both valid and invalid UTF-8,
and it also provides support for Unicode "characters"). But the higher-level
concept of processing grapheme clusters is currently left to external crates.

The 'unicode-segmentation' crate is needed in order to determine the visible
text console screen character slots consumed by the various portions of our
mutt index_format string in its original form, in order for us to substitute
the equivalent number of SPACE chars in the event that we want to suppress a
given field value.