update file: src/main/rust/src/bin/ads-mutt-index-format-custom4.rs
Tighten-up RE_EXPECTED_PATTERN to escape '[' and ']' chars when they appear in
character classes (possitive and negative). This is a weird Rustism regex
thing, and it is not yet clear to me why the regex compiler complained about
some of them the other day but did not complain about these stragglers.