ads-mutt-index-format-custom4.rs: reduce bound ranges
Some of my bounding ranges from commit b0c591a01160 were excessive, but
reasonable considering that they were going from {1,} to {1,something}. Here
we tighten them up, while still allowing for more-than-expected variance.

Keep in mind that even on a large display, the screen space is at a premium,
so the distance used between "fields" of my mutt $index_format variable is
typically exactly two space characters. There are some more optimizations to
be made in that regard, but the tighter we make it, the more fragile we make
it, as well.

For the moment, this seems to be an okay balance (or I'm too tired to
recognize the time lag).